Lua, drag and drop, merging tool
Added scripting tab to tabwidget and a textbox for script log messages.
Started making node merging tool, that will be needed for mirror tool.
User can now drag and drop a jbeam on the editor to open it, or lua
script to run it. Changed nodebeam class to QObject so it can emit and
receive qt signals.